lwip workaround
source:
https://github.com/esp8266/Arduino/commit/efc8dda160c584457804c510c6cbfa
7bb208d8a8

+++ src/ESPAsyncTCP.h
@@ -75,7 +75,7 @@ class AsyncClient {
     int8_t _sent(tcp_pcb* pcb, uint16_t len);
     int8_t _recv(tcp_pcb* pcb, pbuf* pb, int8_t err);
     static int8_t _s_poll(void *arg, struct tcp_pcb *tpcb);
-    static int8_t _s_recv(void *arg, struct tcp_pcb *tpcb, struct pbuf *pb, int8_t err);
+    static int32_t _s_recv(void *arg, struct tcp_pcb *tpcb, struct pbuf *pb, int8_t err);
     static void _s_error(void *arg, int8_t err);
     static int8_t _s_sent(void *arg, struct tcp_pcb *tpcb, uint16_t len);
     static int8_t _s_connected(void* arg, void* tpcb, int8_t err);
